+++ app/serializers/coaching/pledge_serializer.rb
@@ -1,0 +1,2 @@
+class Coaching::PledgeSerializer < PledgeSerializer
+end