+++ app/services/coaching/pledge/filter/base.rb
@@ -1,0 +1,4 @@
+module Coaching::Pledge
+  class Filter::Base < ApplicationFilter
+  end
+end